[avaliacao-3] update productCatalog component
