+++ e-commerce/src/app/home/page.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import BannerCarousel from "./components/banner/banner";
 import { Header } from "../components/header/header";
 import { FooterPage } from "../components/footer/footer.tsx";
-import { ProductCatalog } from "../components/productDetails/productCatalog";
+import { ProductCatalog } from "../components/productDetails/productCatalog.tsx";
 
 export default function HomePage() {
   return (
